refactor(sandbox): local.py 纯函数层拆至 _local_compat（守 1000 行守门）；计划勾选收尾
